fix(HQM-347): Refactor wallet card component
