+++ src/components/wallet-card/index.ts
@@ -1,0 +1,6 @@
+export * from './wallet-card';
+export * from './balance-info-total';
+export * from './balance-info-details';
+export * from './card-name';
+export * from './card-buttons';
+export * from './protection-badge';
